⏪️ UI 분리 작업 롤백
This reverts commit db02a82568b6be1b9c42a71a01d6ec6cf7702c44.
