+++ src/App.tsx
@@ -6,9 +6,19 @@ import {
   EditIcon,
 } from '@chakra-ui/icons';
 import {
+  Alert,
+  AlertDialog,
+  AlertDialogBody,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogOverlay,
+  AlertIcon,
+  AlertTitle,
   Box,
   Button,
   Checkbox,
+  CloseButton,
   Flex,
   FormControl,
   FormLabel,
@@ -17,27 +27,47 @@ import {
   IconButton,
   Input,
   Select,
+  Table,
+  Tbody,
+  Td,
   Text,
+  Th,
+  Thead,
   Tooltip,
+  Tr,
   useToast,
   VStack,
 } from '@chakra-ui/react';
-import { useState } from 'react';
+import { useRef, useState } from 'react';
 
-import MonthView from './components/MonthView.tsx';
-import NotificationAlert from './components/NotificationAlert.tsx';
-import RepeatForm from './components/RepeatForm.tsx';
-import WeekView from './components/WeekView.tsx';
-import { categories, notificationOptions } from './config';
 import { useCalendarView } from './hooks/useCalendarView.ts';
 import { useEventForm } from './hooks/useEventForm.ts';
 import { useEventOperations } from './hooks/useEventOperations.ts';
 import { useNotifications } from './hooks/useNotifications.ts';
 import { useSearch } from './hooks/useSearch.ts';
-import OverlappingAlert from './OverlappingAlert.tsx';
-import { Event, EventForm } from './types';
+import { Event, EventForm, RepeatType } from './types';
+import {
+  formatDate,
+  formatMonth,
+  formatWeek,
+  getEventsForDay,
+  getWeekDates,
+  getWeeksAtMonth,
+} from './utils/dateUtils';
 import { findOverlappingEvents } from './utils/eventOverlap';
 import { getTimeErrorMessage } from './utils/timeValidation';
+
+const categories = ['업무', '개인', '가족', '기타'];
+
+const weekDays = ['일', '월', '화', '수', '목', '금', '토'];
+
+const notificationOptions = [
+  { value: 1, label: '1분 전' },
+  { value: 10, label: '10분 전' },
+  { value: 60, label: '1시간 전' },
+  { value: 120, label: '2시간 전' },
+  { value: 1440, label: '1일 전' },
+];
 
 function App() {
   const {
@@ -83,6 +113,7 @@ function App() {
 
   const [isOverlapDialogOpen, setIsOverlapDialogOpen] = useState(false);
   const [overlappingEvents, setOverlappingEvents] = useState<Event[]>([]);
+  const cancelRef = useRef<HTMLButtonElement>(null);
 
   const toast = useToast();
 
@@ -132,6 +163,128 @@ function App() {
       await saveEvent(eventData);
       resetForm();
     }
+  };
+
+  const renderWeekView = () => {
+    const weekDates = getWeekDates(currentDate);
+    return (
+      <VStack data-testid='week-view' align='stretch' w='full' spacing={4}>
+        <Heading size='md'>{formatWeek(currentDate)}</Heading>
+        <Table variant='simple' w='full'>
+          <Thead>
+            <Tr>
+              {weekDays.map((day) => (
+                <Th key={day} width='14.28%'>
+                  {day}
+                </Th>
+              ))}
+            </Tr>
+          </Thead>
+          <Tbody>
+            <Tr>
+              {weekDates.map((date) => (
+                <Td key={date.toISOString()} height='100px' verticalAlign='top' width='14.28%'>
+                  <Text fontWeight='bold'>{date.getDate()}</Text>
+                  {filteredEvents
+                    .filter((event) => new Date(event.date).toDateString() === date.toDateString())
+                    .map((event) => {
+                      const isNotified = notifiedEvents.includes(event.id);
+                      return (
+                        <Box
+                          key={event.id}
+                          p={1}
+                          my={1}
+                          bg={isNotified ? 'red.100' : 'gray.100'}
+                          borderRadius='md'
+                          fontWeight={isNotified ? 'bold' : 'normal'}
+                          color={isNotified ? 'red.500' : 'inherit'}>
+                          <HStack spacing={1}>
+                            {isNotified && <BellIcon />}
+                            <Text fontSize='sm' noOfLines={1}>
+                              {event.title}
+                            </Text>
+                          </HStack>
+                        </Box>
+                      );
+                    })}
+                </Td>
+              ))}
+            </Tr>
+          </Tbody>
+        </Table>
+      </VStack>
+    );
+  };
+
+  const renderMonthView = () => {
+    const weeks = getWeeksAtMonth(currentDate);
+
+    return (
+      <VStack data-testid='month-view' align='stretch' w='full' spacing={4}>
+        <Heading size='md'>{formatMonth(currentDate)}</Heading>
+        <Table variant='simple' w='full'>
+          <Thead>
+            <Tr>
+              {weekDays.map((day) => (
+                <Th key={day} width='14.28%'>
+                  {day}
+                </Th>
+              ))}
+            </Tr>
+          </Thead>
+          <Tbody>
+            {weeks.map((week, weekIndex) => (
+              <Tr key={weekIndex}>
+                {week.map((day, dayIndex) => {
+                  const dateString = day ? formatDate(currentDate, day) : '';
+                  const holiday = holidays[dateString];
+
+                  return (
+                    <Td
+                      key={dayIndex}
+                      height='100px'
+                      verticalAlign='top'
+                      width='14.28%'
+                      position='relative'>
+                      {day && (
+                        <>
+                          <Text fontWeight='bold'>{day}</Text>
+                          {holiday && (
+                            <Text color='red.500' fontSize='sm'>
+                              {holiday}
+                            </Text>
+                          )}
+                          {getEventsForDay(filteredEvents, day).map((event) => {
+                            const isNotified = notifiedEvents.includes(event.id);
+                            return (
+                              <Box
+                                key={event.id}
+                                p={1}
+                                my={1}
+                                bg={isNotified ? 'red.100' : 'gray.100'}
+                                borderRadius='md'
+                                fontWeight={isNotified ? 'bold' : 'normal'}
+                                color={isNotified ? 'red.500' : 'inherit'}>
+                                <HStack spacing={1}>
+                                  {isNotified && <BellIcon />}
+                                  <Text fontSize='sm' noOfLines={1}>
+                                    {event.title}
+                                  </Text>
+                                </HStack>
+                              </Box>
+                            );
+                          })}
+                        </>
+                      )}
+                    </Td>
+                  );
+                })}
+              </Tr>
+            ))}
+          </Tbody>
+        </Table>
+      </VStack>
+    );
   };
 
   return (
@@ -220,16 +373,38 @@ function App() {
           </FormControl>
 
           {isRepeating && (
-            <RepeatForm
-              {...{
-                repeatType,
-                setRepeatType,
-                repeatInterval,
-                setRepeatInterval,
-                repeatEndDate,
-                setRepeatEndDate,
-              }}
-            />
+            <VStack width='100%'>
+              <FormControl>
+                <FormLabel>반복 유형</FormLabel>
+                <Select
+                  value={repeatType}
+                  onChange={(e) => setRepeatType(e.target.value as RepeatType)}>
+                  <option value='daily'>매일</option>
+                  <option value='weekly'>매주</option>
+                  <option value='monthly'>매월</option>
+                  <option value='yearly'>매년</option>
+                </Select>
+              </FormControl>
+              <HStack width='100%'>
+                <FormControl>
+                  <FormLabel>반복 간격</FormLabel>
+                  <Input
+                    type='number'
+                    value={repeatInterval}
+                    onChange={(e) => setRepeatInterval(Number(e.target.value))}
+                    min={1}
+                  />
+                </FormControl>
+                <FormControl>
+                  <FormLabel>반복 종료일</FormLabel>
+                  <Input
+                    type='date'
+                    value={repeatEndDate}
+                    onChange={(e) => setRepeatEndDate(e.target.value)}
+                  />
+                </FormControl>
+              </HStack>
+            </VStack>
           )}
 
           <Button data-testid='event-submit-button' onClick={addOrUpdateEvent} colorScheme='blue'>
@@ -260,21 +435,8 @@ function App() {
             />
           </HStack>
 
-          {view === 'week' && (
-            <WeekView
-              filteredEvents={filteredEvents}
-              notifiedEvents={notifiedEvents}
-              currentDate={currentDate}
-            />
-          )}
-          {view === 'month' && (
-            <MonthView
-              currentDate={currentDate}
-              filteredEvents={filteredEvents}
-              notifiedEvents={notifiedEvents}
-              holidays={holidays}
-            />
-          )}
+          {view === 'week' && renderWeekView()}
+          {view === 'month' && renderMonthView()}
         </VStack>
 
         <VStack data-testid='event-list' w='500px' h='full' overflowY='auto'>
@@ -348,28 +510,73 @@ function App() {
         </VStack>
       </Flex>
 
-      <OverlappingAlert
-        isOverlapDialogOpen={isOverlapDialogOpen}
-        overlappingEvents={overlappingEvents}
-        notificationTime={notificationTime}
-        saveEvent={saveEvent}
-        setIsOverlapDialogOpen={setIsOverlapDialogOpen}
-        editingEvent={editingEvent}
-        title={title}
-        date={date}
-        startTime={startTime}
-        endTime={endTime}
-        description={description}
-        location={location}
-        category={category}
-        isRepeating={isRepeating}
-        repeatType={repeatType}
-        repeatInterval={repeatInterval}
-        repeatEndDate={repeatEndDate}
-      />
+      <AlertDialog
+        isOpen={isOverlapDialogOpen}
+        leastDestructiveRef={cancelRef}
+        onClose={() => setIsOverlapDialogOpen(false)}>
+        <AlertDialogOverlay>
+          <AlertDialogContent>
+            <AlertDialogHeader fontSize='lg' fontWeight='bold'>
+              일정 겹침 경고
+            </AlertDialogHeader>
+
+            <AlertDialogBody>
+              다음 일정과 겹칩니다:
+              {overlappingEvents.map((event) => (
+                <Text key={event.id}>
+                  {event.title} ({event.date} {event.startTime}-{event.endTime})
+                </Text>
+              ))}
+              계속 진행하시겠습니까?
+            </AlertDialogBody>
+
+            <AlertDialogFooter>
+              <Button ref={cancelRef} onClick={() => setIsOverlapDialogOpen(false)}>
+                취소
+              </Button>
+              <Button
+                colorScheme='red'
+                onClick={() => {
+                  setIsOverlapDialogOpen(false);
+                  saveEvent({
+                    id: editingEvent ? editingEvent.id : undefined,
+                    title,
+                    date,
+                    startTime,
+                    endTime,
+                    description,
+                    location,
+                    category,
+                    repeat: {
+                      type: isRepeating ? repeatType : 'none',
+                      interval: repeatInterval,
+                      endDate: repeatEndDate || undefined,
+                    },
+                    notificationTime,
+                  });
+                }}
+                ml={3}>
+                계속 진행
+              </Button>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialogOverlay>
+      </AlertDialog>
 
       {notifications.length > 0 && (
-        <NotificationAlert notifications={notifications} setNotifications={setNotifications} />
+        <VStack position='fixed' top={4} right={4} spacing={2} align='flex-end'>
+          {notifications.map((notification, index) => (
+            <Alert key={index} status='info' variant='solid' width='auto'>
+              <AlertIcon />
+              <Box flex='1'>
+                <AlertTitle fontSize='sm'>{notification.message}</AlertTitle>
+              </Box>
+              <CloseButton
+                onClick={() => setNotifications((prev) => prev.filter((_, i) => i !== index))}
+              />
+            </Alert>
+          ))}
+        </VStack>
       )}
     </Box>
   );
