--- src/components/NotificationAlert.tsx
@@ -1,27 +1,0 @@
-import { Alert, AlertIcon, AlertTitle, Box, CloseButton, VStack } from '@chakra-ui/react';
-import { Dispatch, SetStateAction } from 'react';
-
-interface NotificationAlertProps {
-  notifications: { id: string; message: string }[];
-  setNotifications: Dispatch<SetStateAction<{ id: string; message: string }[]>>;
-}
-
-const NotificationAlert = ({ notifications, setNotifications }: NotificationAlertProps) => {
-  const handleClose = (index: number) =>
-    setNotifications((prev) => prev.filter((_, i) => i !== index));
-  return (
-    <VStack position='fixed' top={4} right={4} spacing={2} align='flex-end'>
-      {notifications.map((notification, index) => (
-        <Alert key={index} status='info' variant='solid' width='auto'>
-          <AlertIcon />
-          <Box flex='1'>
-            <AlertTitle fontSize='sm'>{notification.message}</AlertTitle>
-          </Box>
-          <CloseButton onClick={() => handleClose(index)} />
-        </Alert>
-      ))}
-    </VStack>
-  );
-};
-
-export default NotificationAlert;
